refactor: split UI port factories by feature
Amp-Thread-ID: https://ampcode.com/threads/T-01a06c03-0b70-73cb-9a3a-6fbc9721e278
